+++ stregsystem/templates/stregsystem/index_sale.html
@@ -29,12 +29,8 @@ for tilsammen {{cost|money}} kr.
 {% endif %}
 {% if caffeine %}
     <br /> <b> Shit der er {{caffeine | floatformat:0}} mg koffein i kroppen</b>
-    {% if caffeine_len >= 1 %}
-        <br />  <b> Det svarer til at drikke {{caffeine_str}} kopper kaffe i streg! </b>
-    {% endif %}
-    {% if caffeine_len == 1 %}
-        <br /> Det svarer til at drikke <b>{{coffeine_str}} kop kaffe!</b>
-    {% endif %}
+    <br /> <b> Det svarer til at drikke {{caffeine | caffeine_emoji_render}} {{ cups|pluralize:'kop kaffe,kopper kaffe i streg!' }}</b>
+{#    Virker det? {{ caffeine_str|caffeine_render }}.#}
 {% endif %}
 {% endautoescape %}
 <br />
