+++ stregsystem/templatetags/stregsystem_extras.py
@@ -1,8 +1,20 @@
 from django import template
 from django.template.loader import get_template
 from django.utils import timezone
+from stregsystem.models import Member
 
 register = template.Library()
+
+@register.filter
+def caffeine_emoji_render(caffeine: int):
+    coffee_str = ""
+
+    from stregsystem.caffeine import CAFFEINE_IN_COFFEE
+    cups = int(caffeine / CAFFEINE_IN_COFFEE)
+    for coffee_cup in range(cups):
+        coffee_str += '☕'  # HTML-code for ☕
+    return coffee_str
+
 
 
 def money(value):
